requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Migrate
psycopg2-binary  # For PostgreSQL
mysql-connector-python # For MySQL
cx_Oracle # For Oracle
python-dotenv # To load environment variables for config
pandas # For data handling in query results
email-validator # For WTForms Email validator
openai # For OpenRouter (OpenAI-compatible API)
Flask-Mail # For sending password reset emails
